% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accumulate_earnings.R
\name{accumulate_earnings}
\alias{accumulate_earnings}
\title{Accumulate earnings}
\usage{
accumulate_earnings(data)
}
\arguments{
\item{data}{tibble with calculated earnings (see `calculate earnings`)}
}
\description{
Accumulate earnings
}
