% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_zero_betting_weeks.R
\name{add_zero_betting_weeks}
\alias{add_zero_betting_weeks}
\title{Add zero betting weeks}
\usage{
add_zero_betting_weeks(data)
}
\arguments{
\item{data}{tibble from calculate_earnings}
}
\description{
Add zero betting weeks
}
