% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arrange_by_odds_group.R
\name{arrange_by_odds_group}
\alias{arrange_by_odds_group}
\title{Arrange by odds group}
\usage{
arrange_by_odds_group(data)
}
\arguments{
\item{data}{tibble obtained from `read_and_prep_data`}
}
\value{
tibble with rows rearranged according to odds group
}
\description{
Arrange by odds group
}
