% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_earnings_for_bets_in_game.R
\name{calculate_earnings_for_bets_in_game}
\alias{calculate_earnings_for_bets_in_game}
\title{Calculate earnings for bets in game}
\usage{
calculate_earnings_for_bets_in_game(data)
}
\arguments{
\item{data}{tibble with bets (see `read_data`)}
}
\description{
Calculate earnings for bets in game
}
