% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_earnings_for_teams.R
\name{calculate_earnings_for_teams}
\alias{calculate_earnings_for_teams}
\title{Calculate earnings for teams}
\usage{
calculate_earnings_for_teams(data)
}
\arguments{
\item{data}{tibble (obtained from df_team_combined())}
}
\description{
Calculate earnings for teams
}
