% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_earnings_grouped_by.R
\name{calculate_earnings_grouped_by}
\alias{calculate_earnings_grouped_by}
\title{Calculate earnings grouped by}
\usage{
calculate_earnings_grouped_by(data, x)
}
\arguments{
\item{data}{tibble (see `read_and_prep_data`)}

\item{x}{variable name (tidy approach, i.e. not strings)}
}
\description{
Calculate earnings grouped by
}
