% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_earnings_grouped_by_stake.R
\name{calculate_earnings_grouped_by_stake}
\alias{calculate_earnings_grouped_by_stake}
\title{Calculate earnings grouped by stake}
\usage{
calculate_earnings_grouped_by_stake(data)
}
\arguments{
\item{data}{tibble (see `read_and_prep_data`)}
}
\description{
Calculate earnings grouped by stake
}
