% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_earnings_grouped_by_team.R
\name{calculate_earnings_grouped_by_team}
\alias{calculate_earnings_grouped_by_team}
\title{Calculate earnings grouped by team}
\usage{
calculate_earnings_grouped_by_team(data, x)
}
\arguments{
\item{data}{variable name (tidy approach)}
}
\description{
Calculate earnings grouped by team
}
