% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_smart_stake.R
\name{calculate_smart_stake}
\alias{calculate_smart_stake}
\title{Calculate smart stake}
\usage{
calculate_smart_stake(odds, stake, eps = 0.5)
}
\arguments{
\item{odds}{odds as numeric value}

\item{stake}{stake as a numeric value}

\item{eps}{epsilon, a numerid value, used fine smart stakes around the stake value (default is 0.5)}
}
\description{
Calculate smart stake
}
