% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/col_prep.R
\name{col_prep}
\alias{col_prep}
\title{Column prep}
\usage{
col_prep(data, breaks_odds, breaks_stake, bound_odds, bound_stake)
}
\arguments{
\item{data}{tibble (see `read_data`)}

\item{breaks_odds}{breaks used to cut odds column}

\item{breaks_stake}{breaks used to cut stake column}

\item{bound_odds}{integer specifying when to bound the odds column (all values above the bound will take the value `bound_odds`)}

\item{bound_stake}{integer specifying when to bound the stake column (all values above the bound will take the value `bound_stake`)}
}
\description{
Column prep
}
