% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cumsum_with_reset.R
\name{cumsum_with_reset}
\alias{cumsum_with_reset}
\title{Cumsum with reset}
\usage{
cumsum_with_reset(x, reset = 0)
}
\arguments{
\item{x}{numeric vector}

\item{reset}{integer specifying reset value (default is zero)}
}
\value{
numeric vector of same length as x
}
\description{
Cumsum with reset
}
