% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_smart_stake.R
\name{floor_to_nearest_half}
\alias{floor_to_nearest_half}
\title{Floor to nearest half}
\usage{
floor_to_nearest_half(x)
}
\arguments{
\item{x}{numeric vector}
}
\description{
Floor to nearest half
}
