% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_accuracy_ci.R
\name{get_accuracy_ci}
\alias{get_accuracy_ci}
\title{Get accuracy confidence interval}
\usage{
get_accuracy_ci(data)
}
\arguments{
\item{data}{tibble with bets (see read_and_prep_data)}
}
\description{
Get accuracy confidence interval
}
