% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_icon_for_info_return.R
\name{get_icon_for_info_return}
\alias{get_icon_for_info_return}
\title{Get icon for info return}
\usage{
get_icon_for_info_return(x)
}
\arguments{
\item{x}{numeric scalar}
}
\description{
Get icon for info return
}
