% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_selectize_choices.R
\name{get_selectize_choices}
\alias{get_selectize_choices}
\title{Get selectize choices}
\usage{
get_selectize_choices(data, x)
}
\arguments{
\item{data}{tibble}

\item{x}{name of variable to pull (must be a column in data)}
}
\description{
Get selectize choices
}
