% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/include_subgames.R
\name{get_subset_and_subgames}
\alias{get_subset_and_subgames}
\title{Get subset and subgames realted to game type filtering}
\usage{
get_subset_and_subgames(data, ...)
}
\arguments{
\item{data}{tibble obtained from `read_and_prep_data`}

\item{...}{filter options}
}
\value{
subset of data with GameType == game_type and the associated doubles and triples included
}
\description{
Get subset and subgames realted to game type filtering
}
