% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_team_names.R
\name{get_team_names}
\alias{get_team_names}
\title{Get team names}
\usage{
get_team_names(data)
}
\arguments{
\item{data}{tibble with historical bets (see `read_data`) having columns HomeTeam and AwayTeam (see map_game_to_*)}
}
\description{
Get team names
}
