% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tournament_names.R
\name{get_tournament_names}
\alias{get_tournament_names}
\title{Get tournament names}
\usage{
get_tournament_names(data)
}
\arguments{
\item{data}{tibble from `read_data`}
}
\description{
Get tournament names
}
