% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_lings_together.R
\name{group_double_and_triple_as_lings}
\alias{group_double_and_triple_as_lings}
\title{Group double and triple as lings}
\usage{
group_double_and_triple_as_lings(x, group)
}
\arguments{
\item{x}{character vector with game types (for instance column GameType from `read_and_prep_data`)}

\item{group}{logical (TRUE/FALSE) indicating whether or not grouping should take place or not.}
}
\description{
Group double and triple as lings
}
