% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_lings_together.R
\name{group_lings_together}
\alias{group_lings_together}
\title{Group lings together}
\usage{
group_lings_together(x, group)
}
\arguments{
\item{x}{character vector with game types (for instance column GameType from `read_and_prep_data`)}

\item{group}{logical (TRUE/FALSE) indicating whether or not grouping should take place or not. 
If FALSE, x is returned}
}
\value{
character vector with game types where all lings have been modified to "X-ling"
}
\description{
Group lings together
}
