% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_game_to_team.R
\name{map_game_to_home_team}
\alias{map_game_to_home_team}
\title{Map game to home team}
\usage{
map_game_to_home_team(x)
}
\arguments{
\item{x}{character string with home team - away team}
}
\value{
character string with home team
}
\description{
Map game to home team
}
