% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_yaxis_to_plot_title.R
\name{map_yaxis_to_plot_title}
\alias{map_yaxis_to_plot_title}
\title{Map yaxis to plot title}
\usage{
map_yaxis_to_plot_title(y)
}
\arguments{
\item{y}{character string with y axis}
}
\description{
Map yaxis to plot title
}
