% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_distribution.R
\name{plot_distribution}
\alias{plot_distribution}
\title{Plot distribution using histogram}
\usage{
plot_distribution(data, var)
}
\arguments{
\item{data}{tibble obtained from `read_and_prep_data`}

\item{var}{character string with column name to select (any from data)}
}
\description{
Plot distribution using histogram
}
