% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_earnings.R
\name{plot_return}
\alias{plot_return}
\title{Plot return}
\usage{
plot_return(data, x)
}
\arguments{
\item{data}{tibble with earnings}

\item{x}{character string with column name to plot returns for}
}
\description{
Plot return
}
