% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_stress.R
\name{select_stress}
\alias{select_stress}
\title{Select stress}
\usage{
select_stress(data, stress)
}
\arguments{
\item{data}{tibble}

\item{stress}{logical, TRUE means stressed columns are shown, FALSE means no stressing}
}
\description{
Select stress
}
