% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/short_tournament_name.R
\name{short_tournament_name}
\alias{short_tournament_name}
\title{Short tournament name}
\usage{
short_tournament_name(x)
}
\arguments{
\item{x}{character vector with tournamnet names}
}
\value{
character vector with abbreviated tournament names
}
\description{
Short tournament name
}
