% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_dropdown_box.R
\name{show_dropdown_box}
\alias{show_dropdown_box}
\title{Show dropdown box}
\usage{
show_dropdown_box(ns, max_stake)
}
\arguments{
\item{ns}{namespace}

\item{max_stake}{maximum stake (integer)}
}
\description{
Show dropdown box
}
