% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_earning_panel.R
\name{show_earning_panel}
\alias{show_earning_panel}
\title{Show earning panel}
\usage{
show_earning_panel(title, outputId, ...)
}
\arguments{
\item{title}{character string with title of panel}

\item{outputId}{outputId of tabel output}

\item{...}{additional arguments applied to shiny::tabPanel}
}
\description{
Show earning panel
}
