% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_accuracy.R
\name{simulate_earnings}
\alias{simulate_earnings}
\title{Simulate earnings}
\usage{
simulate_earnings(data, x, n)
}
\arguments{
\item{data}{tibble (see read_and_prep_data)}

\item{x}{accuracy used to simulate data with (numeric value between 0 and 1)}

\item{n}{number of simulations}
}
\description{
Simulate earnings
}
