% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_accuracy.R
\name{simulate_multiple_earnings}
\alias{simulate_multiple_earnings}
\title{Simulate multiple earnings}
\usage{
simulate_multiple_earnings(data, x, n)
}
\arguments{
\item{data}{tibble (see read_and_prep_data)}

\item{x}{vector with accuracies used to simulate data with (numeric value between 0 and 1)}

\item{n}{number of simulations}
}
\description{
Simulate multiple earnings
}
