% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slice_and_pull.R
\name{slice_and_pull}
\alias{slice_and_pull}
\title{Slice and pull}
\usage{
slice_and_pull(data, i, j)
}
\arguments{
\item{data}{tibble}

\item{i}{integer specifying which row(s) to slice}

\item{j}{variable name (tidy approach) specifying which column to pull}
}
\description{
Slice and pull
}
