% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_distribution.R
\name{summarise_distribution}
\alias{summarise_distribution}
\title{Summarise distribution}
\usage{
summarise_distribution(x)
}
\arguments{
\item{x}{vector of any class}
}
\description{
Summarise distribution
}
