% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_earnings.R
\name{calculate_earnings}
\alias{calculate_earnings}
\title{Calculate earnings}
\usage{
calculate_earnings(
  data,
  var_bets = NULL,
  var_stake = "Stake",
  var_revenue = "Revenue",
  order_by = NULL
)
}
\arguments{
\item{data}{tibble with bets (see `read_data`)}
}
\description{
Calculate earnings
}
