% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/col_prep.R
\name{col_prep}
\alias{col_prep}
\title{Column prep}
\usage{
col_prep(data, breaks)
}
\arguments{
\item{data}{tibble (see `read_data`)}

\item{breaks}{breaks used to cut odds column}
}
\description{
Column prep
}
