% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_by.R
\name{color_by}
\alias{color_by}
\title{Color by}
\usage{
color_by(data, var, colors, na.rm = TRUE)
}
\arguments{
\item{data}{DT::datatable object}

\item{var}{character string with column name to color}

\item{colors}{vector of colors to use}

\item{na.rm}{whether NA values should be removed (default is FALSE)}
}
\description{
Color by
}
