% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_selected_subset.R
\name{get_selected_subset}
\alias{get_selected_subset}
\title{Get selected subset}
\usage{
get_selected_subset(data, ...)
}
\arguments{
\item{data}{tibble obtained from `read_data`}

\item{...}{filter options}
}
\description{
Get selected subset
}
