% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_and_prep_data.R
\name{read_and_prep_data}
\alias{read_and_prep_data}
\title{Read and prep data}
\usage{
read_and_prep_data(breaks)
}
\arguments{
\item{breaks}{breaks used to cut odds column (default is 12-dim sequence from 1 to 14.5)}
}
\description{
Read and prep data
}
